LayoutTests/imported/w3c/web-platform-tests/navigation-api/scroll-behavior/after-transition-timing-expected.txt


FAIL scroll: after-transition should scroll when back completes, just before navigatesuccess assert_equals: expected 0 but got 431
